refactor: extract backend infrastructure, de-duplicate checks, clean a2a_server
server.py: 8,246 -> 7,327 (-919). Four backend clusters moved out of the
god module into their own modules, following the established register() +
re-export pattern:

  ladybug_ops.py   graph store singleton, health/lease probes, investigation
                   and finding mirroring, symbol-index cache, entity lookup
  qdrant_ops.py    client singleton, embedders, payload indexes, upsert,
                   similarity/collection search, cross-encoder rerank
  mnemo_ops.py     Mnemosyne bank access, remember/recall, result coercion
  verdict_ops.py   verdict backend + claim-verdict recording

memcheck/checks/_common.py collects scaffolding shared by contagion,
contradiction and provenance. a2a_server/server.py got the same treatment
the mcp/ side already had: dead code, silent fail-open handlers given
diagnostics, stale Kuzu prose corrected to LadybugDB.

Mutable cache state is deliberately NOT re-exported. `from x import y`
binds by VALUE, so re-exporting ladybug_ops._symbol_index_cache would pin
server.<name> to a None/-1 snapshot while the real cache moves on -- a name
that looks live and is permanently stale. Verified: nothing in the repo
reads it. A comment records why.

Verified across all five suites, unchanged from baseline:
  mcp 488 | mlops 536 | eval 97 | scripts+hooks 556 | a2a 72
Lint clean. Tool invariant 71/71 unique, all re-exported. loci_health still
emits the "kuzu" wire key.
