+++ mcp/memcheck/checks/provenance.py
@@ -26,42 +26,17 @@ import re
 from typing import Callable, Optional
 
 from ..verdict import Verdict, make_signature, new_verdict, redact_excerpt
+from ._common import _default_tokenize, _finding_id
 
 __all__ = ["run_provenance"]
 
 _log = logging.getLogger("memcheck.provenance")
-
-# Mirrors server.py's _TOKEN_RE / _GENERIC_MATCH_TOKENS so the internal
-# fallback tokenizer behaves like the server's when no tokenizer is injected.
-_TOKEN_RE = re.compile(r"[a-z0-9][a-z0-9._:/-]{2,}", re.I)
-_GENERIC_MATCH_TOKENS = {
-    "host", "user", "device", "query", "result", "results", "output", "input", "tool",
-    "found", "seen", "shows", "reported", "detected", "contacted", "event", "events",
-    "record", "records", "row", "rows",
-}
-
-
-def _default_tokenize(text: str) -> set[str]:
-    return {
-        token
-        for token in (m.group(0).lower() for m in _TOKEN_RE.finditer(text or ""))
-        if token not in _GENERIC_MATCH_TOKENS
-    }
 
 
 def _default_lexical_score(a: set[str], b: set[str]) -> float:
     if not a or not b:
         return 0.0
     return len(a & b) / max(1, len(a))
-
-
-def _finding_id(finding: dict, index: int) -> str:
-    """Stable id for a finding: prefer an explicit id, else derive from index."""
-    fid = finding.get("id") or finding.get("finding_id")
-    if fid:
-        return str(fid)
-    inv = finding.get("investigation_id")
-    return f"{inv}:{index}" if inv else f"finding:{index}"
 
 
 def _audit_text(entry: dict) -> str:
